requirements.txt
langchain==0.1.17
langchain-openai==0.1.8
langchain-community==0.0.38
faiss-cpu==1.7.4
pdf2image==1.16.3
pypdf==4.1.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.6.1